Move ScriptHelper into include/, src/ files
